Fix bug: Student ID exceeds the range of int type and CHinese name should be added by fgets instead of scanf

+++ src/Student_Info_Manament_System/include/admin.h
@@ -24,6 +24,6 @@ void delete_student_info(void);
 int find_student_by_username(const char *username, studentInfo *result);
 
 // 根据学号查找学生
-int find_student_by_id(int id, studentInfo *result);
+int find_student_by_id(long long id, studentInfo *result);
 
 #endif // ADMIN_H
